dashboard: update to go111 runtime
The old go 1.9 App Engine runtime is now deprecated. Update to the go 1.11
runtime, which still allows us to use the old App Engine library (removed
in the 1.12 runtime).

Move the app.yaml to the root, as that is preferred by the dev_appserver.py
tooling. <TODO: Do we really want this>

The go111 runtime appears to change the working directory from the app
package directory to the root, so for reading templates, we need to use the
path from the root when we're running in App Engine. This is a nasty hack.

TODO: Nasty directory structure hack

Closes: #1461 ("dashboard: port app to go111/go112 runtime")
Signed-off-by: Andrew Donnellan <ajd@linux.ibm.com>