travis.yml: enable GO111MODULE, update Go versions
Set GO111MODULE="on" in our Travis builds.

Update the Go versions to the latest in the 1.11.x and 1.12.x branches, as
some of our indirect dependencies now declare that they are go 1.12
modules, and prior to 1.11.6(???), a 1.11 toolchain would always fail to
build them. From 1.11.6 onwards, a 1.11 toolchain will attempt to build
them anyway and only complain if the build fails.

Signed-off-by: Andrew Donnellan <ajd@linux.ibm.com>