ci: Stop using GOPATH
Now that we're using modules, we may as well get rid of GOPATH in syz-ci.

Don't build syzkaller inside a GOPATH. Get rid of GOPATH when running
patch/bisect jobs.

TODO: This will break attempts to build old syzkallers, which is something
we do when testing patches or bisecting. I think????

Signed-off-by: Andrew Donnellan <ajd@linux.ibm.com>