pkg/build: separate linux builder for bootloader vs non-bootloader images
Currently, the linux builder uses the create-gce-image.sh script to create
a disk image from a userspace directory and install a bootloader such as
GRUB. As the name of the script suggests, this is useful when using a cloud
platform like GCE, where the platform relies on a bootloader to load the
kernel rather than injecting the kernel directly.

When using the qemu backend on the other hand, it's more efficient to use a
pregenerated root filesystem, and just inject the kernel.

Create some extra configuration fields to specify a path to a root
filesystem image and existing SSH private key file.

Separate the linux builder into "linux", which simply copies the specified
root filesystem, SSH key, and the compiled kernel binary, and "linux_gce",
which uses create-gce-image.sh as is done currently.

Switch all the existing targets to use linux_gce, except ppc64le qemu. This
will break existing users running ppc64le qemu setups, but the main user in
question is myself.

more propagation of rootfs/keypath, TODO: check the bisect stuff
testbuild rootfspath
rootfspath bisect_test

Signed-off-by: Andrew Donnellan <ajd@linux.ibm.com>